+++ SurvivalGame/Assets/Crafting/BuildingCraftBlueprint.cs
@@ -1,0 +1,42 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "NewBuildingBlueprint", menuName = "Game/BuildingBlueprint")]
+public class BuildingBlueprint : CraftBlueprint
+{
+    // public ItemData outputItem;
+    // public int outputQuantity;
+
+    public bool CanBuild(InventorySystem inventory)
+    {
+        // First simple check is to check whenever the required items are available
+        foreach (ItemRequirement requirement in requirements)
+        {
+            if (!inventory.isItemAvailable(new ItemInstance(requirement.item, requirement.quantity)))
+            {
+                Debug.Log($"[BuildingBlueprint] Can't build because {requirement.item.itemName} with quantity {requirement.quantity} is not available");
+                return false;
+            }
+        }
+        return true;
+
+    }
+
+    public bool Build(InventorySystem inventory)
+    {
+        if (CanBuild(inventory))
+        {
+            foreach (ItemRequirement requirement in requirements)
+            {
+                inventory.TryRemoveItem(new ItemInstance(requirement.item, requirement.quantity));
+            }
+
+            // Logic for actual build
+
+            return true;
+        }
+        else
+        {
+            return false;
+        }
+    }
+}